move Rockstar into its own folder
