+++ src/components/Feed/Rockstar/index.js
@@ -1,0 +1,1 @@
+export { default } from './Rockstar'
